Fixed bugs, removed comments

+++ include/Microphone.h
@@ -12,7 +12,7 @@ class Microphone
 public:
   Microphone(String id, int AMP_PIN);
   String getJsonSerializedReadings();
-  unsigned long measureAnalog();
+  int measureAnalog();
 
 private:
   String _id;
